Sperate impl Circuit from test mod into dev mode
