+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -4,9 +4,11 @@
 pub mod bytecode_unroller;
 /// Bytecode circuit
 pub mod circuit;
+#[cfg(any(feature = "test", test, feature = "test-circuits"))]
+mod dev;
 pub(crate) mod param;
 /// Bytecode circuit tester
 #[cfg(any(feature = "test", test))]
-pub mod test;
+mod test;
 #[cfg(any(feature = "test", test, feature = "test-circuits"))]
-pub use test::BytecodeCircuit as TestBytecodeCircuit;
+pub use dev::BytecodeCircuit as TestBytecodeCircuit;
