+++ zkevm-circuits/src/copy_circuit/test.rs
@@ -1,6 +1,7 @@
 #![allow(unused_imports)]
-pub use super::*;
+
 use crate::{
+    copy_circuit::*,
     evm_circuit::{test::rand_bytes, witness::block_convert},
     witness::Block,
 };
@@ -15,70 +16,6 @@ use halo2_proofs::{
     halo2curves::bn256::Fr,
 };
 use mock::{test_ctx::helpers::account_0_code_account_1_no_code, TestContext, MOCK_ACCOUNTS};
-
-impl<F: Field> Circuit<F> for CopyCircuit<F> {
-    type Config = (CopyCircuitConfig<F>, Challenges<Challenge>);
-    type FloorPlanner = SimpleFloorPlanner;
-
-    fn without_witnesses(&self) -> Self {
-        Self::default()
-    }
-
-    fn configure(meta: &mut ConstraintSystem<F>) -> Self::Config {
-        let tx_table = TxTable::construct(meta);
-        let rw_table = RwTable::construct(meta);
-        let bytecode_table = BytecodeTable::construct(meta);
-        let q_enable = meta.fixed_column();
-        let copy_table = CopyTable::construct(meta, q_enable);
-        let challenges = Challenges::construct(meta);
-        let challenge_exprs = challenges.exprs(meta);
-
-        (
-            CopyCircuitConfig::new(
-                meta,
-                CopyCircuitConfigArgs {
-                    tx_table,
-                    rw_table,
-                    bytecode_table,
-                    copy_table,
-                    q_enable,
-                    challenges: challenge_exprs,
-                },
-            ),
-            challenges,
-        )
-    }
-
-    fn synthesize(
-        &self,
-        config: Self::Config,
-        mut layouter: impl Layouter<F>,
-    ) -> Result<(), halo2_proofs::plonk::Error> {
-        let challenge_values = config.1.values(&mut layouter);
-
-        config.0.tx_table.load(
-            &mut layouter,
-            &self.external_data.txs,
-            self.external_data.max_txs,
-            self.external_data.max_calldata,
-            &challenge_values,
-        )?;
-
-        config.0.rw_table.load(
-            &mut layouter,
-            &self.external_data.rws.table_assignments(),
-            self.external_data.max_rws,
-            challenge_values.evm_word(),
-        )?;
-
-        config.0.bytecode_table.load(
-            &mut layouter,
-            self.external_data.bytecodes.values(),
-            &challenge_values,
-        )?;
-        self.synthesize_sub(&config.0, &challenge_values, &mut layouter)
-    }
-}
 
 /// Test copy circuit from copy events and test data
 pub fn test_copy_circuit<F: Field>(
