+++ zkevm-circuits/src/exp_circuit.rs
@@ -1,10 +1,12 @@
 //! Exponentiation verification circuit.
 
+#[cfg(any(feature = "test", test, feature = "test-circuits"))]
+mod dev;
 pub(crate) mod param;
 #[cfg(any(feature = "test", test))]
 mod test;
 #[cfg(any(feature = "test", test, feature = "test-circuits"))]
-pub use test::ExpCircuit as TestExpCircuit;
+pub use dev::ExpCircuit as TestExpCircuit;
 
 use crate::{
     evm_circuit::util::constraint_builder::BaseConstraintBuilder,
