+++ zkevm-circuits/src/pi_circuit.rs
@@ -1,10 +1,12 @@
 //! Public Input Circuit implementation
-
 mod param;
+
+#[cfg(any(feature = "test", test, feature = "test-circuits"))]
+mod dev;
 #[cfg(any(feature = "test", test))]
 mod test;
 #[cfg(any(feature = "test", test, feature = "test-circuits"))]
-pub use test::PiTestCircuit;
+pub use dev::PiTestCircuit;
 
 use eth_types::{
     geth_types::{BlockConstants, Transaction},
